change some tipes from float to double add getTurretYaw

+++ src/main/java/frc/robot/subsystems/IntakeMagazine.java
@@ -6,7 +6,7 @@ public class IntakeMagazine extends SubsystemBase {
     void feedShooter(){
 
     }
-    void setIntakeMotorRPM(float newRPM){
+    void setIntakeMotorRPM(double newRPM){
 
     }
 }
